refactor: 장소 관련 도메인 패키지 구조 변경
