+++ src/main/java/courseitda/place/domain/Place.java
@@ -1,4 +1,4 @@
-package courseitda.workspace.domain;
+package courseitda.place.domain;
 
 import courseitda.common.entity.Timestamp;
 import courseitda.common.exception.BusinessException;
