+++ src/main/java/courseitda/place/infrastructure/JpaPlaceRepository.java
@@ -1,6 +1,6 @@
-package courseitda.workspace.infrastructure;
+package courseitda.place.infrastructure;
 
-import courseitda.workspace.domain.Place;
+import courseitda.place.domain.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaPlaceRepository extends JpaRepository<Place, Long> {
